tools/prioritizers/road_fury/requirements.txt
grpcio>=1.74.0
numpy>=1.26.0
scipy>=1.12.0
torch>=2.0.0
